+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 
 ### Added
 - Add support for managing SSH keys from account console
+- Add Token Introspection URL and Validate Refresh Tokens to OIDC IdP
 
 ### Fixed
 - Correct Linkedin name and logo[RCIAM-1346](https://jira.argo.grnet.gr/browse/RCIAM-1346)